rename integration and sentry.origin, move tests
