+++ dev-packages/browser-integration-tests/suites/tracing/metrics/pageload-measure-spans-domexception-details/init.js
@@ -24,7 +24,7 @@ window.Sentry = Sentry;
 Sentry.init({
   traceLifecycle: 'static',
   dsn: 'https://public@dsn.ingest.sentry.io/1337',
-  integrations: [Sentry.browserTracingIntegration(), Sentry.userTimingSpansIntegration()],
+  integrations: [Sentry.browserTracingIntegration(), Sentry.userTimingIntegration()],
   tracesSampleRate: 1,
 });
