+++ dev-packages/browser-integration-tests/suites/tracing/metrics/pageload-measure-spans/init.js
@@ -12,6 +12,6 @@ window.Sentry = Sentry;
 Sentry.init({
   traceLifecycle: 'static',
   dsn: 'https://public@dsn.ingest.sentry.io/1337',
-  integrations: [Sentry.browserTracingIntegration(), Sentry.userTimingSpansIntegration()],
+  integrations: [Sentry.browserTracingIntegration(), Sentry.userTimingIntegration()],
   tracesSampleRate: 1,
 });
