+++ dev-packages/browser-integration-tests/suites/tracing/user-timing-spans-disabled/init.js
