+++ dev-packages/browser-integration-tests/suites/tracing/user-timing-spans-streamed/init.js
@@ -4,11 +4,7 @@ window.Sentry = Sentry;
 
 Sentry.init({
   dsn: 'https://public@dsn.ingest.sentry.io/1337',
-  integrations: [
-    Sentry.browserTracingIntegration(),
-    Sentry.userTimingSpansIntegration(),
-    Sentry.spanStreamingIntegration(),
-  ],
+  integrations: [Sentry.browserTracingIntegration(), Sentry.userTimingIntegration(), Sentry.spanStreamingIntegration()],
   traceLifecycle: 'stream',
   tracesSampleRate: 1,
 });
