+++ packages/browser/src/index.bundle.tracing.replay.feedback.logs.metrics.ts
@@ -33,7 +33,7 @@ export { setActiveSpanInBrowser } from './tracing/setActiveSpan';
 export { spanStreamingIntegration } from './integrations/spanstreaming';
 export { fetchStreamPerformanceIntegration } from './integrations/fetchStreamPerformance';
 export { webVitalsIntegration } from './integrations/webVitals';
-export { userTimingSpansIntegration } from '@sentry/browser-utils';
+export { userTimingIntegration } from '@sentry/browser-utils';
 
 export { getFeedback, sendFeedback } from '@sentry/feedback';
