+++ packages/browser/src/index.bundle.tracing.replay.feedback.ts
@@ -39,7 +39,7 @@ export { reportPageLoaded } from './tracing/reportPageLoaded';
 export { spanStreamingIntegration } from './integrations/spanstreaming';
 export { fetchStreamPerformanceIntegration } from './integrations/fetchStreamPerformance';
 export { webVitalsIntegration } from './integrations/webVitals';
-export { userTimingSpansIntegration } from '@sentry/browser-utils';
+export { userTimingIntegration } from '@sentry/browser-utils';
 
 export { getFeedback, sendFeedback } from '@sentry/feedback';
