+++ packages/browser/src/index.bundle.tracing.replay.ts
@@ -38,7 +38,7 @@ export { setActiveSpanInBrowser } from './tracing/setActiveSpan';
 export { spanStreamingIntegration } from './integrations/spanstreaming';
 export { fetchStreamPerformanceIntegration } from './integrations/fetchStreamPerformance';
 export { webVitalsIntegration } from './integrations/webVitals';
-export { userTimingSpansIntegration } from '@sentry/browser-utils';
+export { userTimingIntegration } from '@sentry/browser-utils';
 
 export { feedbackIntegrationShim as feedbackAsyncIntegration, feedbackIntegrationShim as feedbackIntegration };
