+++ packages/browser/src/index.ts
@@ -48,7 +48,7 @@ export { setActiveSpanInBrowser } from './tracing/setActiveSpan';
 export { spanStreamingIntegration } from './integrations/spanstreaming';
 export { fetchStreamPerformanceIntegration } from './integrations/fetchStreamPerformance';
 export { webVitalsIntegration } from './integrations/webVitals';
-export { userTimingSpansIntegration } from '@sentry/browser-utils';
+export { userTimingIntegration } from '@sentry/browser-utils';
 
 export type { RequestInstrumentationOptions } from './tracing/request';
 export {
