+++ packages/browser/test/index.bundle.tracing.replay.feedback.test.ts
@@ -5,7 +5,7 @@ import {
   feedbackAsyncIntegration,
   replayIntegration,
   spanStreamingIntegration,
-  userTimingSpansIntegration,
+  userTimingIntegration,
   webVitalsIntegration,
 } from '../src';
 import * as TracingReplayFeedbackBundle from '../src/index.bundle.tracing.replay.feedback';
@@ -17,7 +17,7 @@ describe('index.bundle.tracing.replay.feedback', () => {
     expect(TracingReplayFeedbackBundle.feedbackIntegration).toBe(feedbackAsyncIntegration);
     expect(TracingReplayFeedbackBundle.replayIntegration).toBe(replayIntegration);
     expect(TracingReplayFeedbackBundle.spanStreamingIntegration).toBe(spanStreamingIntegration);
-    expect(TracingReplayFeedbackBundle.userTimingSpansIntegration).toBe(userTimingSpansIntegration);
+    expect(TracingReplayFeedbackBundle.userTimingIntegration).toBe(userTimingIntegration);
     expect(TracingReplayFeedbackBundle.webVitalsIntegration).toBe(webVitalsIntegration);
 
     expect(TracingReplayFeedbackBundle.logger).toBe(loggerShim);
